Complete remaining dual-DB backend project. Not thoroughly tested yet.
AcidState and Postgres are now both potential backends, and the Query
wrapper masks the functionality from the frontend. Also decided that my
error-throwing convention was dumb, and the caller should not have to
keep track of what kinds of errors might get thrown. Instead, the caller
should just have to deal with the reality that any sketchy functions
will be have their results wrapped in Either, with an appropriate error
string if the Left constructor is used. I'm not sure why I didn't decide
to do it this way from the beginning. Oh well, lesson learned.

I have this abusive relationship with myself where I embark on a huge
refactor, and the whole time I'm doing it, I'm like, oh man, this will
make things so much easier in the future! And, like, it does, but the
entire future just consists of huge refactors. Maybe that's productive,
and not abusive.