Provide page templates with a standardized render context
The impetus for this change is that with a growing project it becomes
tedious to individually pass necessary variables to template functions.
Instead, we provide a standardized data structure, RenderContext, which
contains user info, sets of errors/warnings/notices, and HTML that
should only be rendered if no errors are present.

Using the getBaseContext and provideContext functions provided in
View.Util, we can provide a render context for any template function
given we are working in the SessionServerPart monad and have access to a
(not necessarily functioning) database connection.